Fix missing reference error.
Add error condition to check m_FootLocation.
When we destroy the last avatar to spawn a new avatar,
footlocation will set to null and footstep script still try
to spawn footprints. This will throw an error if we don't add
error checking.

Also, add more comments markers for the doxygen.