Add a Mocap noise patterns.
Add Default, Pink and White(Random) noise
patterns to the project.

Created a new animation controller which change
the pattern when the avatar is walking and when
user selected 1 of the 3 noise patterns.

Add functionality to the UI. Now User can select
any noise pattern and avatar will mimic that gait
pattern.

Remark: When exporting the Mocap data retargeting quality
was lowered and Spine3 was discarded. See more details
in import message section under animation.