Baseline Version (Waypoint).
Scene understanding system was struggling a lot
when using in a very large environment. Therefore,
we remove all the componenet related to SU system
and create a waypoint system. User needs to define
the waypoint and the avatar will walk waypoint to
waypoint. This commit has just a simple GUI for that.