Remove coroutine WaitTillAnimationComplete.
Calling coroutine inside update might create unstable behaviour.
Therefore, remove the coroutine based solution and implemented a
lockdown mechanism. I added two events at the start and end position
in animation timeline. When it start, it will lock down the animation
and at the end it will unlock it. We don't go to the next animation
when the previous animation is still in the locked state.