Get noise from file.
Pink and White noises are now reading from a file.
This is just a temporary fix until we figure out how to generate noises inside the code.

Fix an error where new noises are added after the previous set.
To fix this, add a condition to clear up the lists when we hit the cancel button.

Fix the list from growing more than the sample size.

Add a coroutine to export the data file.