Change method to apply gait cycle.
Instead of applying colored noise directly as the speed,
we create desired speed based on animLength / noiseValue.
This will ensure that gait cycle will extend/shrink according
to the noise value. For example, if the noise value is 1.7812,
the next gait cycle will take 1.7812 seconds to complete.

ToDo -> Remove the coroutines in future.